fixed secant transExpr(FuncApp),, now argument will be passed fixed parser fixed translation for constsymbol, but still need to make it more general

+++ src/absyn/Expr.h
@@ -19,7 +19,7 @@ public:
   virtual ~Expr();
 
   void add(Expr* e);
-  Expr* get(int id);
+  Expr* get(size_t id);
   size_t size();
 
 protected:
